+++ static.h
@@ -1,4 +1,4 @@
-const String appVersion = "v1.112";
+const String appVersion = "v1.113";
 #ifndef CUSTOM_APPNAME
    const String appName = "WiTcontroller";
 #else
@@ -8,6 +8,74 @@ const String appVersion = "v1.112";
    #define DEVICE_NAME "WiTcontroller"
 #endif
 
+//
+// DO NOT alter this file
+// 
+// If you need to alter one of the defines in here, do so in config_buttons.h  NOT HERE
+//
+
+// *******************************************************************************************************************
+// *******************************************************************************************************************
+// moved from config_keypad.h
+
+// Keypad + encoder + oLed config
+
+// *******************************************************************************************************************
+// encoder
+
+// default Small ESP32
+#ifndef ROTARY_ENCODER_A_PIN
+    #define ROTARY_ENCODER_A_PIN 12  // to DT
+#endif
+#ifndef ROTARY_ENCODER_B_PIN
+    #define ROTARY_ENCODER_B_PIN 14  // to CLK
+#endif
+#ifndef ROTARY_ENCODER_BUTTON_PIN
+    #define ROTARY_ENCODER_BUTTON_PIN 13  // to SW
+#endif
+
+#ifndef ROTARY_ENCODER_VCC_PIN
+    #define ROTARY_ENCODER_VCC_PIN -1 /* 27 put -1 of Rotary encoder Vcc is connected directly to 3,3V; else you can use declared output pin for powering rotary encoder */
+#endif
+#ifndef ROTARY_ENCODER_STEPS
+    #define ROTARY_ENCODER_STEPS 2 //depending on your encoder - try 1,2 or 4 to get expected behaviour
+#endif
+
+// *******************************************************************************************************************
+// keypad
+
+// 4x3 keypad
+#ifndef ROW_NUM
+    #define ROW_NUM     4
+#endif
+#ifndef COLUMN_NUM
+    #define COLUMN_NUM  3
+#endif
+#ifndef KEYPAD_KEYS
+    #define KEYPAD_KEYS  {'1', '2', '3'}, {'4', '5', '6'}, {'7', '8', '9'}, {'*', '0', '#'}
+#endif
+#ifndef KEYPAD_ROW_PINS
+    #define KEYPAD_ROW_PINS    {19, 18, 17, 16}
+#endif
+#ifndef KEYPAD_COLUMN_PINS
+    #define KEYPAD_COLUMN_PINS { 4, 0, 2}
+#endif
+
+char keys[ROW_NUM][COLUMN_NUM] = {KEYPAD_KEYS};
+byte pin_rows[ROW_NUM]      = KEYPAD_ROW_PINS; // GIOP19, GIOP18, GIOP5, GIOP17 connect to the row pins
+byte pin_column[COLUMN_NUM] = KEYPAD_COLUMN_PINS;   // GIOP16, GIOP4, GIOP0 connect to the column pins
+
+Keypad keypad = Keypad( makeKeymap(keys), pin_rows, pin_column, ROW_NUM, COLUMN_NUM );
+
+#ifndef KEYPAD_DEBOUNCE_TIME
+    #define KEYPAD_DEBOUNCE_TIME 10
+#endif
+#ifndef KEYPAD_HOLD_TIME
+    #define KEYPAD_HOLD_TIME 200
+#endif
+
+// *******************************************************************************************************************
+// *******************************************************************************************************************
 
 #ifndef USE_WIFI_COUNTRY_CODE
    #define USE_WIFI_COUNTRY_CODE false
@@ -222,6 +290,7 @@ const String label_track_power = "TRK";
 
 const int glyph_heartbeat_off = 0x00b7;
 const int glyph_track_power = 0x00eb;
+const int glyph_guest_mode = 0x00ca;
 const int glyph_speed_step = 0x00d6;
 // const int glyph_direction_forward = 0x0070;
 // const int glyph_direction_reverse = 0x006d;
@@ -1186,4 +1255,50 @@ const char ssidPasswordBlankChar = 164;
 
 #ifndef CONSIST_RELEASE_BY_INDEX
    #define CONSIST_RELEASE_BY_INDEX true
+#endif
+
+// ***************************************************
+// function search
+
+#ifndef MAX_FUNCTION_SEARCH
+   #define MAX_FUNCTION_SEARCH 4
+#endif
+
+#ifndef FUNCTION_SEARCH_LABELS
+   #define FUNCTION_SEARCH_LABELS      { {"Horn","Whistle"}, {"Horn",""}, {"Whistle",""}, {"Brake",""} }
+#endif
+
+#ifndef FUNCTION_SEARCH_IDS
+   #define FUNCTION_SEARCH_IDS {HORN_OR_WHISTLE_SEARCH, HORN_SEARCH, WHISTLE_SEARCH, BRAKE_SEARCH} 
+#endif
+
+#ifndef FUNCTION_SEARCH_RANDOM_MATCH
+   #define FUNCTION_SEARCH_RANDOM_MATCH false
+#endif
+
+// ***************************************************
+// guest mode
+
+#ifndef GUEST_MODE_ENABLED
+   #define GUEST_MODE_ENABLED false
+#endif
+
+#ifndef GUEST_MODE_PIN_1
+   #define GUEST_MODE_PIN_1 41
+#endif
+
+#ifndef GUEST_MODE_PIN_2
+   #define GUEST_MODE_PIN_2 42
+#endif
+
+#ifndef GUEST_MODE_LED_PIN
+   #define GUEST_MODE_LED_PIN -1
+#endif
+
+#ifndef GUEST_MODE_HOLD_DURATION
+   #define GUEST_MODE_HOLD_DURATION 1000
+#endif
+
+#ifndef GESTURE_PARTNER_WINDOW
+   #define GESTURE_PARTNER_WINDOW 250
 #endif